requirements.txt
teradatasql
sqlalchemy
psycopg2-binary
pytest
python-dotenv
